Protected Route

+++ Backend/controller/goalController.js
@@ -1,8 +1,11 @@
 const asyncHandler = require("express-async-handler");
 const Goal = require("../models/goalModel");
+const User = require("../models/userModel");
 
 const getGoals = async (req, res) => {
-  const goals = await Goal.find();
+  console.log(req.headers.authorization);
+  const goals = await Goal.find({ userID: req.user._id });
+  // console.log('user::', req.user.id);
   res.status(200).json(goals);
 };
 
@@ -14,6 +17,7 @@ const createGoals = asyncHandler(async (req, res) => {
 
   const goal = await Goal.create({
     text: req.body.text,
+    user: req.user.id,
   });
   res.status(200).json(goal);
 });
@@ -26,6 +30,18 @@ const updateGoals = asyncHandler(async (req, res) => {
     throw new Error("Goal not found");
   }
 
+  const user = await User.findById(req.user.id);
+  // Check for user
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+  // making sure the loggedIn user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
+  }
+
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
   });
@@ -34,14 +50,27 @@ const updateGoals = asyncHandler(async (req, res) => {
 });
 
 const deleteGoals = asyncHandler(async (req, res) => {
-    const goal = await Goal.findById(req.params.id)
-    if (!goal) {
-        res.status(400);
-        throw new Error("Goal not found");
-      }
-    await goal.deleteOne()
-    // await goal.remove()
-    res.status(200).json({ id: req.params.id })
+  const goal = await Goal.findById(req.params.id);
+  if (!goal) {
+    res.status(400);
+    throw new Error("Goal not found");
+  }
+
+  const user = await User.findById(req.user.id);
+  // Check for user
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+  // making sure the loggedIn user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
+  }
+
+  await goal.deleteOne();
+  // await goal.remove()
+  res.status(200).json({ id: req.params.id });
 });
 
 module.exports = {
